+++ QtLing/QtLing.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 4.6.2, 2018-07-16T10:50:00. -->
+<!-- Written by QtCreator 4.6.2, 2018-07-16T14:05:07. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
